Revert "Refactor configurations"
